将代码与测试分离 使用gradle packJar命令打包 打包后只有一个jar  包含desktop、ios、iosmoe、android 将根据平台类型使用反射加载 删掉了一些gradle任务
