+++ android/src/main/java/net/mwplay/nativefont/test/AndroidLauncher.java
@@ -1,0 +1,16 @@
+package net.mwplay.nativefont.test;
+
+import android.os.Bundle;
+
+import com.badlogic.gdx.backends.android.AndroidApplication;
+import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
+
+
+public class AndroidLauncher extends AndroidApplication {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        AndroidApplicationConfiguration config = new AndroidApplicationConfiguration();
+        initialize(new FontTestGame(), config);
+    }
+}
