+++ core/src/core/java/net/mwplay/nativefont/NativeFontListener.java
@@ -1,0 +1,11 @@
+package net.mwplay.nativefont;
+
+import com.badlogic.gdx.graphics.Pixmap;
+
+/**
+ * Created by tian on 2016/10/2.
+ */
+
+public interface NativeFontListener {
+    Pixmap getFontPixmap(String str, NativeFontPaint freePaint);
+}
