--- core/src/net/mwplay/nativefont/NativeFontListener.java
@@ -1,11 +1,0 @@
-package net.mwplay.nativefont;
-
-import com.badlogic.gdx.graphics.Pixmap;
-
-/**
- * Created by tian on 2016/10/2.
- */
-
-public interface NativeFontListener {
-    Pixmap getFontPixmap(String str, NativeFontPaint freePaint);
-}
