+++ desktop/src/main/java/net/mwplay/nativefont/test/DesktopLauncher.java
@@ -1,0 +1,14 @@
+package net.mwplay.nativefont.test;
+
+import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
+import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
+
+
+public class DesktopLauncher {
+    public static void main(String[] arg) {
+        LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+        config.width = 480;
+        config.height = 800;
+        new LwjglApplication(new FontTestGame(), config);
+    }
+}
