+++ ios-moe/proguard.append.cfg
@@ -1,3 +1,4 @@
 
 -keep class com.badlogic.** { *; }
 -keep enum com.badlogic.** { *; }
+-keep class net.mwplay.** { *; }
