+++ ios-moe/src/main/java/net/mwplay/nativefont/test/IOSMoeLauncher.java
@@ -1,0 +1,28 @@
+package net.mwplay.nativefont.test;
+
+
+import com.badlogic.gdx.backends.iosmoe.IOSApplication;
+import com.badlogic.gdx.backends.iosmoe.IOSApplicationConfiguration;
+
+import org.moe.natj.general.Pointer;
+
+import apple.uikit.c.UIKit;
+
+
+public class IOSMoeLauncher extends IOSApplication.Delegate {
+
+    protected IOSMoeLauncher(Pointer peer) {
+        super(peer);
+    }
+
+    @Override
+    protected IOSApplication createApplication() {
+        IOSApplicationConfiguration config = new IOSApplicationConfiguration();
+        config.useAccelerometer = false;
+        return new IOSApplication(new FontTestGame(), config);
+    }
+
+    public static void main(String[] argv) {
+        UIKit.UIApplicationMain(0, null, null, IOSMoeLauncher.class.getName());
+    }
+}
